Updated VSCode to 1.2.1 and added some extensions I want installed with it.

+++ vscode/tools/chocolateyInstall.ps1
@@ -1,7 +1,7 @@
 $packageName = "vscode"
 $installerType = "EXE"
 $installerArgs = "/SILENT /MERGETASKS=!runCode,addtopath,addcontextmenufolders"
-$url = "https://az764295.vo.msecnd.net/stable/809e7b30e928e0c430141b3e6abf1f63aaf55589/VSCodeSetup-stable.exe"
+$url = "https://az764295.vo.msecnd.net/stable/fe7f407b95b7f78405846188259504b34ef72761/VSCodeSetup-stable.exe"
 
 if ($psISE) {
     Import-Module -name "$env:ChocolateyInstall\chocolateyinstall\helpers\chocolateyInstaller.psm1"
@@ -9,8 +9,23 @@ if ($psISE) {
 
 Install-ChocolateyPackage $packageName $installerType $installerArgs $url
 
-if (Test-ProcessAdminRights) {
-    Remove-Item "$($env:PUBLIC)\Desktop\Visual Studio Code.lnk" -Force
-} else {
-    Start-ChocolateyProcessAsAdmin "Remove-Item '$($env:PUBLIC)\Desktop\Visual Studio Code.lnk' -Force"
+if (Test-Path "$($env:PUBLIC)\Desktop\Visual Studio Code.lnk") {
+    if (Test-ProcessAdminRights) {
+        Remove-Item "$($env:PUBLIC)\Desktop\Visual Studio Code.lnk" -Force
+    } else {
+        Start-ChocolateyProcessAsAdmin "Remove-Item '$($env:PUBLIC)\Desktop\Visual Studio Code.lnk' -Force"
+    }
 }
+
+Write-Output "Install some VS Code extensions..."
+
+$code = "$env:PF32\Microsoft VS Code\bin\code.cmd"
+
+Start-Process -FilePath $code -ArgumentList "--install-extension ms-vscode.csharp" -NoNewWindow -Wait
+Start-Process -FilePath $code -ArgumentList "--install-extension donjayamanne.python" -NoNewWindow -Wait
+Start-Process -FilePath $code -ArgumentList "--install-extension msjsdiag.debugger-for-chrome" -NoNewWindow -Wait
+Start-Process -FilePath $code -ArgumentList "--install-extension samverschueren.yo" -NoNewWindow -Wait
+Start-Process -FilePath $code -ArgumentList "--install-extension rprouse.theme-obsidian" -NoNewWindow -Wait
+Start-Process -FilePath $code -ArgumentList "--install-extension PeterJausovec.vscode-docker" -NoNewWindow -Wait
+Start-Process -FilePath $code -ArgumentList "--install-extension Pendrica.Chef" -NoNewWindow -Wait
+
